+++ app/services/line_bot_service.rb
@@ -1,4 +1,5 @@
 class LineBotService < BaseService
+  include LineFormat
   include Conversion
 
   attr_accessor :graph, :google, :user, :request
@@ -41,25 +42,6 @@ class LineBotService < BaseService
     return return_msg
   end
 
-  def text_format return_msg
-    { type: 'text',
-      text: return_msg }
-  end
-
-  def carousel_format columns
-    { type: "template",
-      altText: I18n.t('carousel.text'),
-      template: {
-        type: "carousel",
-        columns: columns }}
-  end
-
-  def button_format text, link
-    { type: "uri",
-      label: text,
-      uri: link }
-  end
-
   def varify_signature
     body = request.body.read
     signature = request.env['HTTP_X_LINE_SIGNATURE']
@@ -69,27 +51,5 @@ class LineBotService < BaseService
   def find_line_user id
     User.create!(line_user_id: id) if !User.exists?(line_user_id: id)
     self.user = User.find_by(line_user_id: id)
-  end
-
-  def carousel_options results
-    columns = []
-
-    results.each do |result|
-      r = facebook_response(result)
-      r.text = set_text(r, 'line')
-
-      actions = []
-      actions << button_format(I18n.t('button.official'), safe_url(r.link_url))
-      actions << button_format(I18n.t('button.location'), safe_url(google.get_map_link(r.lat, r.lng, r.name, r.street)))
-      actions << button_format(I18n.t('button.related_comment'), safe_url(google.get_google_search(r.name)))
-
-      columns << {
-        thumbnailImageUrl: r.image_url,
-        title: r.name,
-        text: r.text,
-        actions: actions }
-    end
-    
-    return columns
   end
 end
